All classes are defined in one file.
I attempted to follow MVC guidelines, this is the cart model file.  It defines an abstract shopping cart class, then extends that class to two concrete classes.  The first uses PHP_SESSION variables to hold the shopping cart when the user isn't logged in.  The second uses a MySQL database to store the shopping cart items when the user is logged in.  The working implementation can be seen at: php.tadaatiedye.com.